backend/requirements-ci.txt
# B&R Capital Dashboard - CI/Testing Requirements
# Lightweight dependencies for CI pipeline (excludes heavy ML libraries)
# Full requirements.txt used for production deployment

# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
alembic>=1.13.1

# Redis Caching
redis>=5.0.1

# WebSocket Support
websockets>=12.0
python-socketio>=5.11.0

# Email Service
aiosmtplib>=3.0.1
email-validator>=2.1.0
jinja2>=3.1.3

# Data Processing (lightweight - excludes torch, tensorflow)
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dateutil>=2.8.2
numpy>=1.26.3
pandas>=2.2.0

# Excel Extraction
pyxlsb>=1.0.10
openpyxl>=3.1.0
msal>=1.24.0
structlog>=23.1.0
aiohttp>=3.8.0
apscheduler>=3.10.0

# Authentication & Security
PyJWT[crypto]>=2.8.0,<3.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0,<5.0.0
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.1
loguru>=0.7.2
tenacity>=8.2.3

# Monitoring & Metrics
prometheus-client>=0.19.0
psutil>=5.9.8

# API Documentation
openapi-schema-pydantic>=1.2.4

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
pytest-timeout>=2.2.0
aiosqlite>=0.19.0

# Development/CI Tools
ruff>=0.4.0
black>=24.1.1
isort>=5.13.2
mypy>=1.8.0
flake8>=7.0.0

# Security Scanning
bandit>=1.7.0
safety>=2.3.0
